# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponce:
    properties:
      message:
        type: string
    type: object
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  todo.List:
    properties:
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  todo.User:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for TodoList Application
  title: Todo App API
  version: "1.0"
paths:
  /api/lists:
    post:
      consumes:
      - application/json
      description: create todo list
      operationId: create-list
      parameters:
      - description: list info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo.List'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponce'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponce'
      summary: Create todo list
      tags:
      - lists
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponce'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponce'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponce'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponce'
      summary: SignUp
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"